build: tag build alpine docker container
with either a passed through tag or the current commit

+++ docd/alpine.sh
@@ -1,18 +1,22 @@
 #!/usr/bin/env sh
 
-export NAME=docd
-export VERSION=alpine
+main() {
+  export NAME=docd
+  export VERSION=alpine
+  
+  if [ "$#" -ne 1 ];
+    then echo "usage: ${0} <version>"
+    exit 1
+  fi
 
-function die {
-  echo 1>&2 $*
-  exit 1
+
+  echo "Building ${NAME} for ${VERSION}..."
+
+  GOOS=linux GOARCH=amd64 go build -o $NAME || exit 1
+  cp docd $VERSION
+  cd $VERSION || exit 1
+  docker build -t "$NAME:$1" -t "$NAME:latest" . || exit 1
+  rm docd  
 }
 
-echo "Building ${NAME} for ${VERSION}..."
-
-GOOS=linux GOARCH=amd64 go build -o $NAME || exit 1
-cp docd $VERSION
-cd $VERSION
-docker build -t $NAME . || exit 1
-rm docd
-
+main "$@"
